pyproject.toml
[project]
name = "ixtalk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kurigram>=2.2.7",
    "python-dotenv>=1.1.1",
    "tgcrypto>=1.2.5",
]
